userscorrigido

+++ app/Http/Controllers/UserController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Support\Facades\Auth;
 use App\Models\User;
 use Illuminate\Http\Request;
 
@@ -52,6 +53,13 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        if (
+            auth()->user()->role !== 'super_admin' &&
+            auth()->id() !== $user->id
+        ) {
+            abort(403, 'Acesso negado');
+        }
+
         return view('users.edit', compact('user'));
     }
 
@@ -60,6 +68,14 @@ class UserController extends Controller
      */
     public function update(Request $request, User $user)
     {
+
+        if (
+            auth()->user()->role !== 'super_admin' &&
+            auth()->id() !== $user->id
+        ) {
+            abort(403, 'Acesso negado');
+        }
+
         $data = $request->validate([
             'name'            => ['required', 'string', 'max:255'],
             'cpf'             => ['required', 'string', 'max:14', 'unique:users,cpf,' . $user->id],
@@ -87,7 +103,21 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        if (
+            auth()->user()->role !== 'super_admin' &&
+            auth()->id() !== $user->id
+        ) {
+            abort(403, 'Acesso negado');
+        }
+
+        $isOwnAccount = auth()->id() === $user->id;
+
         $user->delete();
+
+        if ($isOwnAccount) {
+            Auth::logout();
+            return redirect('/login')->with('success', 'Conta excluída com sucesso!');
+        }
 
         return redirect()
             ->route('users.index')
